requirements.txt
black
Django==3.2.5
djangorestframework==3.12.4
psycopg2==2.9.1
model-bakery
isort